Handle timeouts when using the history purge endpoint
This adds a new "purge_request_timeout" configuration setting, which allows
configuring a request timeout for requests to the hitory purge API endpoint
*only*. These timeouts are converted into PurgeTimeout exceptions, which get
handled by allowing the process to move over to the next room being purged.
Synapse will still be trying to complete the purge, even if the request timed
out, so a timeout here does not mean that things failed.