Added experimental concurrent reindex command
Added a new function in the Database class to perform the
reindexation of the huge tables. This fuctions creates a concurrent
secondary index using CREATE INDEX CONCURRENTLY. When the new index
is ready, it performs the DROP of the older index and replaces it
with the newest.

This operation has 2 advantages respect the reindex based on the
REINDEX TABLE operation:

* It doesn't lock the table so the service is not affected
* It is much faster than the REINDEX

In the other hand:

* Creation of the new INDEX could fail (catched as exepction)
* More disk usage is required because we are duplicating the index

Related: #10

Committed suggested modifications during the PR review:

* __execute() is now a privated method
* Fixed the descrition of reindex command
* Uses formatted string instance of .format()