Use all_rooms instead of public_rooms
If possible we should cover all the rooms. Using the HTTP API it is not
possible to see all the rooms with the current code, but using PostgreSQL
directly is recommended anyway.  At some point we may want to use the HTTP API
in the same way Riot does to fetch bridged network rooms, which would allow
for obtaining more rooms than just fetching the main public rooms directory.